+++ tests/test_config.py
@@ -16,7 +16,9 @@ import neural_lam.config as nlconfig
 )
 def test_config_serialization(state_weighting_config):
     c = nlconfig.NeuralLAMConfig(
-        datastore=nlconfig.DatastoreSelection(kind="mdp", config_path=""),
+        datastores={
+            "main": nlconfig.DatastoreSelection(kind="mdp", config_path="")
+        },
         training=nlconfig.TrainingConfig(
             state_feature_weighting=state_weighting_config
         ),
@@ -27,22 +29,26 @@ def test_config_serialization(state_weighting_config):
 
 
 yaml_training_defaults = """
-datastore:
-  kind: mdp
-  config_path: ""
+datastores:
+  main:
+    kind: mdp
+    config_path: ""
 """
 
 default_config = nlconfig.NeuralLAMConfig(
-    datastore=nlconfig.DatastoreSelection(kind="mdp", config_path=""),
+    datastores={
+        "main": nlconfig.DatastoreSelection(kind="mdp", config_path="")
+    },
     training=nlconfig.TrainingConfig(
         state_feature_weighting=nlconfig.UniformFeatureWeighting()
     ),
 )
 
 yaml_training_manual_weights = """
-datastore:
-  kind: mdp
-  config_path: ""
+datastores:
+  main:
+    kind: mdp
+    config_path: ""
 training:
   state_feature_weighting:
     __config_class__: ManualStateFeatureWeighting
@@ -52,7 +58,9 @@ training:
 """
 
 manual_weights_config = nlconfig.NeuralLAMConfig(
-    datastore=nlconfig.DatastoreSelection(kind="mdp", config_path=""),
+    datastores={
+        "main": nlconfig.DatastoreSelection(kind="mdp", config_path="")
+    },
     training=nlconfig.TrainingConfig(
         state_feature_weighting=nlconfig.ManualStateFeatureWeighting(
             weights=dict(u100m=1.0, v100m=1.0)